Add feature of sub creates
Therefore we need to update the db sheame so it can support a hirachie of creates.
Therefor we create a possibilety to refernce a parent create
`parent_id`.

And for performance reasons we also store the `level` of a create which
is equal to the number of times you can go to the parent create untill
you end up at the root level '0'.

We also store a back reference to all children (ifinity levels) of a create
so we can do selects/statistics on them in the database quickly.

So if you have the create structure: fruit -> apple -> foo
Than this will result in the following table structure:

||||||||||||||||||||||||||||||||||||||||||||||||||
|        | id | parent_id | level | children_ids |
| fruite | 10 | 0         | 0     | 20,30        |
| apple  | 20 | 10        | 1     | 30           |
| foo    | 30 | 20        | 2     | NULL         |
||||||||||||||||||||||||||||||||||||||||||||||||||